+++ concordium-sdk-examples/src/main/java/com/concordium/sdk/examples/contractexample/parameters/WCCDTransferParam.java
@@ -7,7 +7,7 @@ import com.concordium.sdk.transactions.smartcontracts.parameters.ListParam;
 import java.util.List;
 
 /**
- * Wrapper class for {@link ListParam} enforcing the correct {@link Transfer} id for cis2-wccd contract.
+ * Wrapper class for {@link ListParam} enforcing the correct {@link Cis2Transfer} id for cis2-wccd contract.
  */
 public class WCCDTransferParam extends ListParam {
     public WCCDTransferParam(Schema schema, ReceiveName receiveName, List<WCCDTransfer> transfers) {