+++ concordium-sdk-examples/src/main/java/com/concordium/sdk/examples/contractexample/wccd/Cis2WCCD.java
@@ -5,6 +5,7 @@ import com.concordium.sdk.Connection;
 import com.concordium.sdk.crypto.ed25519.ED25519SecretKey;
 import com.concordium.sdk.requests.AccountQuery;
 import com.concordium.sdk.requests.BlockQuery;
+import com.concordium.sdk.responses.blockitemstatus.FinalizedBlockItem;
 import com.concordium.sdk.responses.modulelist.ModuleRef;
 import com.concordium.sdk.transactions.*;
 import com.concordium.sdk.transactions.smartcontracts.SchemaParameter;
@@ -12,7 +13,6 @@ import com.concordium.sdk.types.AccountAddress;
 import com.concordium.sdk.types.ContractAddress;
 import com.concordium.sdk.types.Nonce;
 import com.concordium.sdk.types.UInt64;
-import lombok.SneakyThrows;
 import lombok.var;
 import picocli.CommandLine;
 
@@ -33,7 +33,7 @@ public class Cis2WCCD implements Callable<Integer> {
     private static final Expiry EXPIRY = Expiry.createNew().addMinutes(5);
 
     private static final TransactionSigner SIGNER = TransactionSigner.from(
-            SignerEntry.from(Index.from(0), Index.from(0), // TODO dummy key
+            SignerEntry.from(Index.from(0), Index.from(0),
                     ED25519SecretKey.from("56f60de843790c308dac2d59a5eec9f6b1649513f827e5a13d7038accfe31784")) // Dummy key
     );
 
@@ -138,9 +138,8 @@ public class Cis2WCCD implements Callable<Integer> {
                 .build();
         Hash txHash = client.sendTransaction(initContractTransaction);
         System.out.println("Submitted transaction for " + this.methodName  + " with hash: " + txHash);
-        sleep();
-        Hash blockHash = client.getBlockItemStatus(txHash).getFinalizedBlockItem().get().getBlockHash();
-        System.out.println("Transaction finalized in block with hash: " + blockHash);
+        FinalizedBlockItem finalizedTransaction = client.waitUntilFinalized(txHash, timeout);
+        System.out.println("Transaction finalized in block with hash: " + finalizedTransaction.getBlockHash());
     }
 
     private void handleMethod(ClientV2 client, Nonce nonce, SchemaParameter parameter) {
@@ -155,21 +154,10 @@ public class Cis2WCCD implements Callable<Integer> {
                 .build();
         Hash txHash = client.sendTransaction(transaction);
         System.out.println("Submitted transaction for " + this.methodName  + " with hash: " + txHash);
-        sleep();
-        Hash blockHash = client.getBlockItemStatus(txHash).getFinalizedBlockItem().get().getBlockHash();
-        System.out.println("Transaction finalized in block with hash: " + blockHash);
+        FinalizedBlockItem finalizedTransaction = client.waitUntilFinalized(txHash, timeout);
+        System.out.println("Transaction finalized in block with hash: " + finalizedTransaction.getBlockHash());
     }
 
-    @SneakyThrows
-    private void sleep() {
-        for (int i = 0; i < 200; i++) {
-            System.out.print(".");
-            if (i % 50 == 0) {
-                System.out.println("\n");
-            }
-            Thread.sleep(30);
-        }
-    }
 
     public static void main(String[] args) {
         int exitCode = new CommandLine(new Cis2WCCD()).execute(args);